removed twilio responses from profile page and converted it to a class component
