+++ client/src/pages/Profile.js
@@ -5,7 +5,6 @@ import withAuth from './../components/withAuth';
 import API from './../utils/API';
 import LocationList from '../components/LocationList';
 import contentEditable from '../components/ContentEditable';
-import TwilioResponses from '../components/TwilioResponses';
 
 import Sidebar from './../components/Sidebar/Sidebar.jsx';
 import dashboardRoutes from "./../dashboard/routes/dashboard.jsx";
@@ -17,10 +16,6 @@ class Profile extends Component {
     username: "",
     email: "",
     locations: [],
-    surResValid: "",
-    surResInvalid: "",
-    comResValid: "",
-    comResInvalid: ""
   };
 
   componentDidMount() {
@@ -29,10 +24,6 @@ class Profile extends Component {
         username: res.data.username,
         email: res.data.email,
         locations: res.data.locations,
-        // surResValid: res.data.twilioResponses.surResValid || "",
-        // surResInvalid: res.data.twilioResponses.surResInvalid || "",
-        // comResValid: res.data.twilioResponses.comResValid || "",
-        // comResInvalid: res.data.twilioResponses.comResInvalid || ""
       })
     });
 
@@ -50,20 +41,7 @@ class Profile extends Component {
 
   }
 
-  handleInputChange = event => {
-    const { name, value } = event.target;
-    this.setState({
-      [name]: value
-    });
-  };
 
-  handleTwilioResponsesSubmit = event => {
-    event.preventDefault();
-    API
-      .submitTwilioResponses(this.state.surResValid, this.state.surResInvalid, this.state.comResValid, this.state.comResInvalid)
-      .then(res => console.log(res.data))
-      .catch(err => console.log(err));
-  }
 
 
 
@@ -75,45 +53,34 @@ class Profile extends Component {
       <div className="wrapper">
         <Sidebar {...this.props} routes={dashboardRoutes} />
         <div className="main-panel" ref="mainPanel">
-        <div style={{marginBottom:"50px"}}><Header dashColor={"black"} {...this.props} /></div>
-        <hr/>
-      <div className="container Profile">
-        <h3>Profile page</h3>
-          <div className = "container">
-              <div className = "row">
+          <div style={{ marginBottom: "50px" }}><Header dashColor={"black"} {...this.props} /></div>
+          <hr />
+          <div className="container Profile">
+            <h3>Profile page</h3>
+            <div className="container">
+              <div className="row">
                 <p>Username:</p> <EditableUserName value={this.state.username} id="username" />
               </div>
-              <div className = "row">
-                <p>Email address:</p> <EditableEmail value={this.state.email} id="email"/>
+              <div className="row">
+                <p>Email address:</p> <EditableEmail value={this.state.email} id="email" />
               </div>
-          </div>
-
-          <div>        
-            <div>
-              <LocationList 
-                userid={this.props.user.id}
-                history={this.props.history}
-              />
             </div>
-            <Link to="/addlocation">Add a location</Link>&nbsp;&nbsp;&nbsp;&nbsp;
+
+            <div>
+              <div>
+                <LocationList
+                  userid={this.props.user.id}
+                  history={this.props.history}
+                />
+              </div>
+              <Link to="/addlocation">Add a location</Link>&nbsp;&nbsp;&nbsp;&nbsp;
             <Link to="/importlocations/">Import locations</Link>&nbsp;&nbsp;&nbsp;&nbsp;
             <Link to="/">Go home</Link>
+            </div>
+
+
           </div>
-
-        <div>
-            <TwilioResponses 
-              handleInputChange={this.handleInputChange}
-              handleTwilioResponsesSubmit = {this.handleTwilioResponsesSubmit}
-              responses={this.state.twilioResponses}
-              surResValid={this.state.surResValid}
-              surResInvalid={this.state.surResInvalid}
-              comResValid={this.state.comResValid}
-              comResInvalid={this.state.comResInvalid}
-            />
-      </div>
-
-      </div>
-      </div>
+        </div>
       </div>
     )
   }
